requirements.txt
flask
flask-cors
requests
beautifulsoup4
openai
python-dotenv
gunicorn
